Recompilacion previa a Copa 22/12/2019

+++ blogictierFAP/src/main/java/freewill/nextgen/blts/data/ClassicShowEntity.java
@@ -46,7 +46,7 @@ public class ClassicShowEntity implements Serializable, Cloneable {
     private float PVTotal=0;
     private float puntuacionTotal=0;
 	private int clasificacionFinal=0; //Puesto en el que quedó el patinador en resultado final
-	private float sumaPonderada = 0;
+	private double sumaPonderada = 0;
 	
     /**
      * Default constructor. 
@@ -316,11 +316,11 @@ public class ClassicShowEntity implements Serializable, Cloneable {
 		this.puntuacionTotal = puntuacionTotal;
 	}
 
-	public float getSumaPonderada() {
+	public double getSumaPonderada() {
 		return sumaPonderada;
 	}
 
-	public void setSumaPonderada(float sumaPonderada) {
+	public void setSumaPonderada(double sumaPonderada) {
 		this.sumaPonderada = sumaPonderada;
 	}
 }
